+++ lib/info_about_triangle.dart
