+++ lib/painters/Painters/RightTrianglePainter.dart
@@ -1,8 +1,9 @@
-import 'package:flutter/material.dart';
-import '../../constants.dart';
-
-import '../../triangleBrain.dart';
 import 'dart:math' as math;
+
+import 'package:flutter/material.dart';
+
+import '../../constants.dart';
+import '../../triangle_brain.dart';
 
 class RightTrianglePainter extends CustomPainter {
   TriangleModel triangle = TriangleModel();
