+++ lib/widgets/decimal_bar.dart
